refactored and created components for both date and time filter option sets and imported them into viewDrafts component
